+++ oxe-api/resource/campaign/add_campaign_addresses.py
@@ -8,7 +8,7 @@ from sqlalchemy.orm.exc import NoResultFound
 from decorator.catch_exception import catch_exception
 from decorator.log_request import log_request
 from decorator.verify_admin_access import verify_admin_access
-from utils.re import has_mail_format
+from utils.regex import has_mail_format
 from exception.object_not_found import ObjectNotFound