+++ oxe-api/resource/user/update_user.py
@@ -7,7 +7,7 @@ from webargs import fields
 from decorator.catch_exception import catch_exception
 from decorator.log_request import log_request
 from decorator.verify_admin_access import verify_admin_access
-from utils.re import has_mail_format
+from utils.regex import has_mail_format
 
 
 class UpdateUser(MethodResource, Resource):